add commandtable and init test

+++ db/cmd/string.go
@@ -7,6 +7,8 @@ import (
 )
 
 func Set(db *db.DataBase, cmd [][]byte) resp.RedisData {
+
+	// 进行输入类型检查
 	cmdName := strings.ToLower(string(cmd[0]))
 	if cmdName != "set" {
 		return resp.MakeErrorData("Server error")
@@ -16,11 +18,17 @@ func Set(db *db.DataBase, cmd [][]byte) resp.RedisData {
 		return resp.MakeErrorData("error: commands is invalid")
 	}
 
-	// 检查 key 是否被设置为 ttl，如果设置需要删除
+	// 进行类型检查，会自动检查过期选项
+	r, ok := CheckOldType(db, string(cmd[1]), STRING)
+	if ok == MISMATCH {
+		return r
+	}
 
-	db.RemoveTTL(string(cmd[1]))
-
+	// 键值对设置
 	db.SetKey(string(cmd[1]), string(cmd[2]))
+
+	// 重置 TTL
+	db.RemoveTTL(string(cmd[1]))
 
 	return resp.MakeStringData("OK")
 }
@@ -41,14 +49,24 @@ func Get(db *db.DataBase, cmd [][]byte) resp.RedisData {
 	value, ok := db.GetKey(string(cmd[1]))
 	if !ok {
 		return resp.MakeStringData("nil")
-
 	}
 
 	strVal, ok := value.(string)
 	if !ok {
-		println("convert error")
+		return resp.MakeErrorData("WRONGTYPE Operation against a key holding the wrong kind of value")
 	}
-	byteVal := []byte(strVal)
 
-	return resp.MakeBulkData(byteVal)
+	return resp.MakeBulkData([]byte(strVal))
+}
+
+func StrLen(db *db.DataBase, cmd [][]byte) resp.RedisData {
+
+	return resp.MakeIntData(-1)
+}
+
+func RegisterStringCommands() {
+
+	RegisterCommand("set", Set)
+	RegisterCommand("get", Get)
+
 }
